@@ -17,7 +17,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   get: ^4.6.1
   shared_preferences: ^2.0.9
-  kakao_flutter_sdk: ^0.9.0
+  kakao_flutter_sdk: ^1.2.2
   uuid: 3.0.5
   flutter_facebook_auth: ^4.1.1
   flutter_web_auth: ^0.3.1
