@@ -175,7 +175,7 @@ class findPw_view extends GetView<HomeController>{
                                 SizedBox(
                                   height: 48,
                                   width: double.infinity,
-                                  child: RaisedButton(
+                                  child: ElevatedButton(
                                     onPressed: () {
                                       FindPwViewModel.findPw(_id, _phone);
                                     },
@@ -184,20 +184,22 @@ class findPw_view extends GetView<HomeController>{
                                       style: TextStyle(
                                           fontSize: 16, color: Colors.white),
                                     ),
-                                    color: const Color.fromARGB(
-                                        255, 0, 104, 166),
-                                    shape: RoundedRectangleBorder(
-                                      side: const BorderSide(
-                                          color: Color.fromARGB(
-                                              255, 0, 104, 166)),
-                                      borderRadius: BorderRadius.circular(10),
+                                    style: ElevatedButton.styleFrom(
+                                      shape: RoundedRectangleBorder(
+                                        side: const BorderSide(
+                                            color: Color.fromARGB(
+                                                255, 0, 104, 166)),
+                                        borderRadius: BorderRadius.circular(10),
+                                      ),
+                                      backgroundColor: const Color.fromARGB(
+                                          255, 0, 104, 166),
                                     ),
                                   ),
                                 ),
                                 SizedBox(
                                   height: 48,
                                   width: double.infinity,
-                                  child: RaisedButton(
+                                  child: ElevatedButton(
                                     onPressed: () {
                                       Navigator.pop(context);
                                     },
@@ -207,13 +209,15 @@ class findPw_view extends GetView<HomeController>{
                                           fontSize: 16, color: Color.fromARGB(
                                           255, 0, 104, 166)),
                                     ),
-                                    color: const Color.fromARGB(
-                                        255, 255, 255, 255),
-                                    shape: RoundedRectangleBorder(
-                                      side: const BorderSide(
-                                          color: Color.fromARGB(
-                                              255, 0, 104, 166)),
-                                      borderRadius: BorderRadius.circular(10),
+                                    style: ElevatedButton.styleFrom(
+                                      shape: RoundedRectangleBorder(
+                                        side: const BorderSide(
+                                            color: Color.fromARGB(
+                                                255, 0, 104, 166)),
+                                        borderRadius: BorderRadius.circular(10),
+                                      ),
+                                      backgroundColor: const Color.fromARGB(
+                                          255, 0, 104, 166),
                                     ),
                                   ),
                                 )
