@@ -112,7 +112,7 @@ class intro_view extends GetView<HomeController> {
                     child: SizedBox(
                       height: 48,
                       width: double.infinity,
-                      child: RaisedButton(
+                      child: ElevatedButton(
                         onPressed: () {
                           Get.to(() => newMember1_view());
                         },
@@ -120,10 +120,15 @@ class intro_view extends GetView<HomeController> {
                           "다음으로",
                           style: TextStyle(fontSize: 16, color: Colors.white),
                         ),
-                        color: const Color.fromARGB(97, 255, 255, 255),
-                        shape: RoundedRectangleBorder(
-                          side: const BorderSide(color: Colors.white),
-                          borderRadius: BorderRadius.circular(10),
+                        style: ElevatedButton.styleFrom(
+                          shape: RoundedRectangleBorder(
+                            side: const BorderSide(
+                                color: Color.fromARGB(
+                                    255, 0, 104, 166)),
+                            borderRadius: BorderRadius.circular(10),
+                          ),
+                          backgroundColor: const Color.fromARGB(
+                              255, 0, 104, 166),
                         ),
                       ),
                     ),
