@@ -15,8 +15,8 @@ class login_view extends GetView<LoginController>{
   String id = '';
   String passwd = '';
 
-  var _id = TextEditingController();
-  var _passwd = TextEditingController();
+  final _id = TextEditingController();
+  final _passwd = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
@@ -221,7 +221,7 @@ class login_view extends GetView<LoginController>{
                                         child: SizedBox(
                                           height: 48,
                                           width: double.infinity,
-                                          child: RaisedButton(
+                                          child: ElevatedButton(
                                             onPressed: () {
                                               LoginViewModel.login(_id, _passwd,);
                                             },
@@ -230,11 +230,13 @@ class login_view extends GetView<LoginController>{
                                               style:
                                               TextStyle(fontSize: 16, color: Colors.white),
                                             ),
-                                            color: const Color.fromARGB(97, 255, 255, 255),
-                                            shape: RoundedRectangleBorder(
-                                              side: const BorderSide(color: Colors.white),
-                                              borderRadius: BorderRadius.circular(10),
-                                            ),
+                                            style: ElevatedButton.styleFrom(
+                                              backgroundColor: const Color.fromARGB(97, 255, 255, 255),
+                                              shape: RoundedRectangleBorder(
+                                                side: const BorderSide(color: Colors.white),
+                                                borderRadius: BorderRadius.circular(10),
+                                              ),
+                                            )
                                           ),
                                         ),
                                       ),
