@@ -70,7 +70,7 @@ class navigator_view extends GetView<NavigatorController>{
                               SizedBox(
                                 height: 48,
                                 width: double.infinity,
-                                child: RaisedButton(
+                                child:ElevatedButton (
                                   onPressed: () {
                                     Get.to(() => intro_view());
                                   },
@@ -79,19 +79,22 @@ class navigator_view extends GetView<NavigatorController>{
                                     style: TextStyle(
                                         fontSize: 16, color: Colors.white),
                                   ),
-                                  color: const Color.fromARGB(
-                                      97, 255, 255, 255),
-                                  shape: RoundedRectangleBorder(
-                                    side: const BorderSide(
-                                        color: Colors.white),
-                                    borderRadius: BorderRadius.circular(10),
+                                  style: ElevatedButton.styleFrom(
+                                    shape: RoundedRectangleBorder(
+                                      side: const BorderSide(
+                                          color: Color.fromARGB(
+                                              255, 0, 104, 166)),
+                                      borderRadius: BorderRadius.circular(10),
+                                    ),
+                                    backgroundColor: const Color.fromARGB(
+                                        255, 0, 104, 166),
                                   ),
                                 ),
                               ),
                               SizedBox(
                                 height: 48,
                                 width: double.infinity,
-                                child: RaisedButton(
+                                child: ElevatedButton(
                                   onPressed: () {
                                     Get.to(() => login_view());
                                   },
@@ -100,12 +103,15 @@ class navigator_view extends GetView<NavigatorController>{
                                     style: TextStyle(
                                         fontSize: 16, color: Colors.white),
                                   ),
-                                  color: const Color.fromARGB(
-                                      97, 255, 255, 255),
-                                  shape: RoundedRectangleBorder(
-                                    side: const BorderSide(
-                                        color: Colors.white),
-                                    borderRadius: BorderRadius.circular(10),
+                                  style: ElevatedButton.styleFrom(
+                                    shape: RoundedRectangleBorder(
+                                      side: const BorderSide(
+                                          color: Color.fromARGB(
+                                              255, 0, 104, 166)),
+                                      borderRadius: BorderRadius.circular(10),
+                                    ),
+                                    backgroundColor: const Color.fromARGB(
+                                        255, 0, 104, 166),
                                   ),
                                 ),
                               )
