@@ -120,7 +120,7 @@ class newMember1_view extends GetView<NewMemberController>{
                     child: SizedBox(
                       height: 48,
                       width: double.infinity,
-                      child: RaisedButton(
+                      child: ElevatedButton(
                         onPressed: () {
                           if (NewMemberViewModel.isStudentChecked == true ||
                               NewMemberViewModel.isTeacherChecked == true) {
@@ -138,10 +138,15 @@ class newMember1_view extends GetView<NewMemberController>{
                           "다음으로",
                           style: TextStyle(fontSize: 16, color: Colors.white),
                         ),
-                        color: const Color.fromARGB(97, 255, 255, 255),
-                        shape: RoundedRectangleBorder(
-                          side: const BorderSide(color: Colors.white),
-                          borderRadius: BorderRadius.circular(10),
+                        style: ElevatedButton.styleFrom(
+                          shape: RoundedRectangleBorder(
+                            side: const BorderSide(
+                                color: Color.fromARGB(
+                                    255, 0, 104, 166)),
+                            borderRadius: BorderRadius.circular(10),
+                          ),
+                          backgroundColor: const Color.fromARGB(
+                              255, 0, 104, 166),
                         ),
                       ),
                     ),
