@@ -251,7 +251,7 @@ class newMember5_view extends GetView<NewMemberController> {
                             SizedBox(
                               height: 48,
                               width: double.infinity,
-                              child: RaisedButton(
+                              child: ElevatedButton(
                                 onPressed: () {
                                   NewMemberViewModel.requestJoinProcess();
                                 },
@@ -260,20 +260,22 @@ class newMember5_view extends GetView<NewMemberController> {
                                   style: TextStyle(
                                       fontSize: 16, color: Colors.white),
                                 ),
-                                color: const Color.fromARGB(
-                                    255, 0, 104, 166),
-                                shape: RoundedRectangleBorder(
-                                  side: const BorderSide(
-                                      color: Color.fromARGB(
-                                          255, 0, 104, 166)),
-                                  borderRadius: BorderRadius.circular(10),
-                                ),
+                                style: ElevatedButton.styleFrom(
+                                  backgroundColor: const Color.fromARGB(
+                                      255, 0, 104, 166),
+                                  shape: RoundedRectangleBorder(
+                                    side: const BorderSide(
+                                        color: Color.fromARGB(
+                                            255, 0, 104, 166)),
+                                    borderRadius: BorderRadius.circular(10),
+                                  ),
+                                )
                               ),
                             ),
                             SizedBox(
                               height: 48,
                               width: double.infinity,
-                              child: RaisedButton(
+                              child: ElevatedButton(
                                 onPressed: () {
                                   Navigator.pop(context);
                                 },
@@ -283,14 +285,16 @@ class newMember5_view extends GetView<NewMemberController> {
                                       fontSize: 16, color: Color.fromARGB(
                                       255, 0, 104, 166)),
                                 ),
-                                color: const Color.fromARGB(
-                                    255, 255, 255, 255),
-                                shape: RoundedRectangleBorder(
-                                  side: const BorderSide(
-                                      color: Color.fromARGB(
-                                          255, 0, 104, 166)),
-                                  borderRadius: BorderRadius.circular(10),
-                                ),
+                                style: ElevatedButton.styleFrom(
+                                  backgroundColor: const Color.fromARGB(
+                                      255, 255, 255, 255),
+                                  shape: RoundedRectangleBorder(
+                                    side: const BorderSide(
+                                        color: Color.fromARGB(
+                                            255, 0, 104, 166)),
+                                    borderRadius: BorderRadius.circular(10),
+                                  ),
+                                )
                               ),
                             )
                           ],
