@@ -32,7 +32,7 @@ class QuitWidget extends StatelessWidget {
         ],
       ),
       actions: <Widget>[
-        FlatButton(
+        ElevatedButton(
           child: const Text("확인"),
           onPressed: () {
             SystemChannels.platform.invokeMethod('SystemNavigator.pop');
