@@ -41,7 +41,7 @@ class SendWidget extends StatelessWidget {
         ],
       ),
       actions: <Widget>[
-        FlatButton(
+        ElevatedButton(
           child: const Text("확인"),
           onPressed: () {
             Get.back();
